Refactor include structure to best practice
Maybe not quite ideal yet; we may still end up moving all internal
headers into src in some way.
