+++ Makefile
@@ -16,6 +16,7 @@ BUILD_DIR = build
 TESTS_DIR = tests
 INCLUDE_DIR = include
 SUBMODULE_DIR = external
+INTERNAL_INCLUDE_DIR = $(INCLUDE_DIR)/internal
 C_SRC_DIR = ${SRC_DIR}/c
 ASM_SRC_DIR = ${SRC_DIR}/asm
 
@@ -66,13 +67,13 @@ external/googletest/lib/libgtest_main.a external/googletest/lib/libgtest.a:
 
 $(BUILD_DIR)/tests.o: external/googletest/lib/libgtest_main.a external/googletest/lib/libgtest.a \
 	                    $(TESTS) $(TESTS_DIR)/compare_buffer_functions.cpp $(LIBS_DIR)/test_$(LIBNAME).a 
-	$(CC) $(CFLAGS) -I$(INCLUDE_DIR) -I$(GTEST_INCLUDE_DIR) $^ -o$@
+	$(CC) $(CFLAGS) -I$(INTERNAL_INCLUDE_DIR) -I$(GTEST_INCLUDE_DIR) $^ -o$@
 
 clean:
 	rm $(BUILD_DIR)/*
 	rm $(LIBS_DIR)/*
 
-deep-clean:
+clean-all:
 	rm $(BUILD_DIR)/*
 	rm $(LIBS_DIR)/*
 	(cd $(SUBMODULE_DIR) && make clean)
