+++ include/internal/aolc/compare_buffer_functions.h
