+++ include/internal/aolc/string_extensions.h
