+++ tests/test_linkages.cpp
@@ -1,8 +1,8 @@
-#include "internal/_test_string.h"
+#include "aolc/_test_string.h"
 #include <string.h>
 #include <signal.h>
 
-#include "internal/compare_buffer_functions.h"
+#include "aolc/compare_buffer_functions.h"
 
 #include "gtest/gtest.h"
