+++ tests/test_memset.cpp
@@ -1,7 +1,7 @@
-#include "internal/_test_string.h"
+#include "aolc/_test_string.h"
 #include <string.h>
 
-#include "internal/compare_buffer_functions.h"
+#include "aolc/compare_buffer_functions.h"
 #include "gtest/gtest.h"
 
 TEST(memset, Basic)
